Add policy configuration for handling errors in PostMasterFailoverProcesses
In some cases a failure in PostMasterFailoverProcesses may leave
Orchestrator's infrastructure in an unhealthy state.

By default Orchestrator will report such failover handling to be
successful, when this may not be considered correct.

Add a new configuration setting PostMasterFailoverProcessesFailureConsideredFailure
which if true will treat the failover as unsuccessful and report
this back to the caller.